requirements.txt
psycopg2-binary
pytest
dbt-postgres
apache-airflow
